chore(markdown): Refactor MentionsPlugin
