+++ markdown/src/main/java/it/niedermann/android/markdown/markwon/plugins/mentions/PotentialAvatarSpan.java
@@ -1,0 +1,6 @@
+package it.niedermann.android.markdown.markwon.plugins.mentions;
+
+import androidx.annotation.NonNull;
+
+record PotentialAvatarSpan(@NonNull String userId, @NonNull String url) {
+}
