+++ markdown/src/main/java/it/niedermann/android/markdown/markwon/plugins/mentions/PotentialDisplayNameSpan.java
@@ -1,0 +1,6 @@
+package it.niedermann.android.markdown.markwon.plugins.mentions;
+
+import androidx.annotation.NonNull;
+
+record PotentialDisplayNameSpan(@NonNull String userId) {
+}
